Ensure cannot search without filters
In order to do this I had to refactor the form to store filters in state
under its own key so that I could inspect it. Also had to stop setting
default values in state in order to inspect if any filters selected.

Added a getFilter method in order to set default values for form.