Abstract out legacy/current attachment logic
Historically the booking in tool uploaded attachments such a Proof of Delivery and Packing Lists to a folder on a PHP server. It would then add the filename to a comma separated list in a table. When rewriting this functionality we wanted to maintain backwards compatibility with this method but also store new uploads in a table using the paperclip gem.

Alex originally implemented this in GoodsReceivedNotice only for Packing Lists but we are going to want to reuse this logic for proof of delivery uploads. I have moved the packing list specific methods out of GoodsReceivedNotice into library classes that are generic save for the PackingListAttachments which implements some packing list specific stuff.

Luckily Alex has tested the majority of this logic via packing list and GRN feature specs but I have implemented some unit level tests too.