Prevent importing barcodes with rejected chars
Theres several instances now of Service Desk having to adjust barcodes
because they have been imported with either leading or trailing quotes
('").

One solution was to strip them but I am worried that if one day valid
barcodes do have quotes then we will silently break them.

The other (and my selected solution) was to just say it's invalid. It
forces whomever is importing the barcodes to check the data and correct
their own mistake instead of pushing the task to Service Desk.